renaming to hamm
