+++ automation/src/test/scala/org/broadinstitute/workbench/hamm/automation/AutomationTestConfig.scala
