+++ core/src/main/scala/org/broadinstitute/workbench/hamm/db/db.scala
@@ -1,4 +1,4 @@
-package org.broadinstitute.workbench.ccm
+package org.broadinstitute.workbench.hamm
 
 import java.time.Instant
 import java.util.UUID
