+++ core/src/test/scala/org/broadinstitute/workbench/hamm/db/DummyDbTransactor.scala
@@ -1,4 +1,4 @@
-package org.broadinstitute.workbench.ccm
+package org.broadinstitute.workbench.hamm
 package db
 
 import cats.effect.{ContextShift, IO}
