+++ core/src/test/scala/org/broadinstitute/workbench/hamm/db/WorkflowCostDAOSqlSpec.scala
@@ -1,10 +1,10 @@
-package org.broadinstitute.workbench.ccm.db
+package org.broadinstitute.workbench.hamm.db
 
 import cats.effect.IO
 import doobie.implicits._
 import doobie.specs2._
-import org.broadinstitute.workbench.ccm.Generators._
-import org.broadinstitute.workbench.ccm.db.WorkflowCostDAO._
+import org.broadinstitute.workbench.hamm.Generators._
+import org.broadinstitute.workbench.hamm.db.WorkflowCostDAO._
 import org.specs2.mutable.Specification
 
 import scala.concurrent.ExecutionContext
